Change all assert lambdas to take an implicit receiver
ex:
```
assert(foo).isNotNull { it.isEqualTo("bar") }
```
is now
```
assert(foo).isNotNull { isEqualTo("bar") }
```

It's not very useful to give these receivers an explicit name and there
is now a warning when `it` is nested.

Unfortunately this is a breaking change with no automated migration
path. We are not yet 1.0 though.