Change all, and assertAll to take a list of assertion lambdas
Removed the concept of 'soft assertions' that don't throw, instead all
assertions throw, and `all` and `assertAll` loop through lambda with a
try/catch. This removes the foot-gun of an assertion maybe running the
rest of the body when it fails, an allows props like `index` and `key`
to return a result and fail with an exception instead of having to take
a lambda.

This has been done in a way to not cause existing code to fail to
compile. Instead, you may not see all exceptions thrown. Deprecated
functions have been kept to guide migration.