Change the signature of `isEqualTo` from Any? to T (#238)
You may think this wouldn't make a difference since T is covariant so
you can pass any anything as they share the same super-type `Any?`.
However, there _is_ a case where this is different and it has do with
kotliln's type inference on literal numbers.

If you have a literal number like `1` in the source code, kotlin will
try to infer it's type (byte, short, int, long) based on where it's
used. Ex: `val b: Byte = 1`. If it can't infer, it'll default to int.

By changing to use a generic type param instead of `Any?` for the
expected value, kotlin will now be able to infer the type.

Before this change:

```
assertThat(1L).isEqualTo(1)
```
would fail with `expected:<1[]> but was:<1[L]>`.

After this change:

```
assertThat(1L).isEqualTo(1)
```
passes.

I think this is worth doing because you don't have to worry about
specifying the type for the number literal in the common case that you
provide it directly. This matches other uses of them pretty well.