Fix NoSuchElementException thrown by containsSubList
Fix an edge case where calling containsSubList() throws a NoSuchElementException instead of an AssertionError when the sublist is empty but the actual list is non-empty.

This was covered by unit tests, but they only verified that an exception was thrown, not the type of the exception, and they specified that this case should not be allowed. This commit changes the behavior to allow [] as a sublist of any list, which matches the behavior of startsWith() and my intuition.