Simplify generics of isInstanceOf
'S' is only used to limit the receiver which means we can replace it with its bound ('Any') and let the 'out' variance support any subtype (such as 'Assert<String>') remain valid.

As far as I can tell this second generic was introduced as part of b817dc3c8b7915b3bbb147a4d82144dd377a478a, where 'S' is actually used to hold an instance. But when the lambda was eventually removed the extra generic stuck around until today.